+++ src/util/index.ts
@@ -143,3 +143,41 @@ export function requestAnimationFrame(cb: Function) {
       cb();
     });
 }
+
+export function compact(value: [] | object): [] | object {
+  if (Array.isArray(value)) {
+    return value.filter(Boolean);
+  }
+  if (!isObject(value)) return value;
+  let keys = Object.keys(value);
+  if (keys.length <= 0) return value;
+  let res = {};
+  for (let i = 0, len = keys.length; i < len; i++) {
+    let item = value[keys[i]];
+    if (Array.isArray(item) || isObject(item)) {
+      res[keys[i]] = compact(item);
+    } else if (!!item) {
+      res[keys[i]] = item;
+    }
+  }
+  return res;
+}
+
+export const compactUndefined = (value: [] | object): [] | object => {
+  if (Array.isArray(value)) {
+    return value.filter((v) => v === void 0);
+  }
+  if (!isObject(value)) return value;
+  let keys = Object.keys(value);
+  if (keys.length <= 0) return value;
+  let res = {};
+  for (let i = 0, len = keys.length; i < len; i++) {
+    let item = value[keys[i]];
+    if (Array.isArray(item) || isObject(item)) {
+      res[keys[i]] = compact(item);
+    } else if (item !== void 0) {
+      res[keys[i]] = item;
+    }
+  }
+  return res;
+};