+++ config/dev.js
@@ -1,9 +1,6 @@
+/* eslint-disable import/no-commonjs */
 module.exports = {
   env: {
-    NODE_ENV: '"development"'
+    NODE_ENV: '"development"',
   },
-  defineConstants: {
-  },
-  mini: {},
-  h5: {}
-}
+};