+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Features
 
+- **cli:** add taro platform change config ([2ecc23e](https://github.com/VFiee/github-test/commit/2ecc23efb1f7eed1c5ea6ce47bf060c67fc6bf50))
 - **component:** add transition component ([d907385](https://github.com/VFiee/github-test/commit/d907385d0f19dcbb964aa0b38e3621406b195885))
 - **component:** 增加 options,使组件受外部 classNameclassName 影响 ([f27efb7](https://github.com/VFiee/github-test/commit/f27efb7bade400072e2e4c1b4d74bdcfcb74cf38))
 - **component:** 完善 image 组件,新增 icon 组件 ([f0f8a51](https://github.com/VFiee/github-test/commit/f0f8a513419cc267b778238abc9c4bff03d1bca4))